feat: Allow unicode fields to work with the path lexer. So we can support different languages object fields.
